fix: address review comment and move util into utils and also the tests. Rename log info
