+++ providers/aws/utils/utils.go
@@ -91,3 +91,10 @@ func GetPriceMap(pricingOutput *pricing.GetProductsOutput, field string) (map[st
 
 	return priceMap, nil
 }
+
+func Int64PtrToFloat64(i *int64) float64 {
+    if i == nil {
+        return 0.0  // or any default value you prefer
+    }
+    return float64(*i)
+}
