+++ providers/aws/utils/utils_test.go
@@ -234,3 +234,24 @@ func TestGetPriceMap_NoPricingOutput(t *testing.T) {
 		t.Errorf("Expected an empty priceMap, but got %v", priceMap)
 	}
 }
+
+func TestInt64PtrToFloat64_ValidInput(t *testing.T) {
+    var number int64 = 1
+    pointer := &number
+
+    returnValue := Int64PtrToFloat64(pointer)
+    var expected float64 = 1.0
+    if returnValue != expected {
+        t.Errorf("Expected return value: %f, but got: %f", expected, returnValue)
+    }
+}
+
+func TestInt64PtrToFloat64_NilInput(t *testing.T) {
+    // nil input
+    returnValue := Int64PtrToFloat64(nil)
+    var expected float64 = 0.0
+    if returnValue != expected {
+        t.Errorf("Expected return value: %f, but got: %f", expected, returnValue)
+    }
+}
+
